Move BadComponent Out of Main Components Directory
